refactor: Rename utils/message/{hasOuterWhitespace,entryInformation}.ts, fix editablePattern comments
