+++ translate/src/utils/message/entryInformation.ts
